requirements.txt
polars>=1.0.0,<2.0.0
oracledb>=2.1.2,<3.0.0
fastapi>=0.111.0,<1.0.0
uvicorn>=0.30.1,<1.0.0
python-docx
openpyxl
xlsxwriter
pydantic>=2.0.0,<3.0.0
python-dotenv
keyring
python-multipart
pyarrow>=14.0.0
